Fix parsing of scontrol output in slurm epilog collector
The previous implementation assumed that the output of scontrol would
consist only of key-value pairs of the form `key=value`. As pointed out
by Raphael Kleinemuehl, this is not necessarily the case. Also, cases
where `key=something=blah` where also parsed wrong, leading to `(key,
something)`, ignoring `=blah`. This is fixed in this commit. Strings not
containing an `=` are ignored and `key=v1=v2` is correctly parsed to
`(key, v1=v2)`.