Add option to allow for empty values in sacct output
Up to now empty values in the sacct output were filtered out.
This commit adds a new option to the key config, which allows empty
fields to pass the filtering. Per default, empty fields are still not
allowed. The option has to be enabled on a per-key basis.

To better handle the key config, a new `KeyConfig` struct is introduced,
which replaces the former tuple of key name and type, i.e.
`(String, ParsableType)`.