Add semver tags and edge tag for docker containers
For each tag following the `vX.Y.Z` syntax that is pushed to GitHub, a
docker container with the tag `X.Y.Z` is created. In addition, the tags
`X.Y` and `X` are also created or updated. Furthermore, the `latest` tag
will point to container of the latest tag, and the `edge` tag will point
to the container corresponding to the latest commit of the `main`
branch.