fix(ci): Split build/push and updating latest symlink into two jobs
The shalzz/zola-deploy-action action uses a docker container to build
the webpage. This messes up the permissions of the output files (i.e.
they belong to root) and therefore the cannot be overwritten later
(without sudo).

When we push a tag we need to update the symlink in the `gh-pages`
branch. By putting this part into a new job we don't have issues
overwriting the output from the webpage build when switching branches.

To keep things in sync the same setup is used for the pyauditor docs,
even though we don't necessarily need it here.